+++ src/get/parents.js
@@ -12,7 +12,11 @@ export default function({context} = {}) {
 
   while (element) {
     list.push(element);
-    element = element.parentElement;
+    // IE does know support parentElement on SVGElement
+    element = element.parentNode;
+    if (element.nodeType !== Node.ELEMENT_NODE) {
+      element = null;
+    }
   }
 
   return list;