+++ src/is/focus-relevant.supports.js
@@ -6,8 +6,11 @@ import canFocusFieldset from '../supports/focus-fieldset';
 import canFocusFlexboxContainer from '../supports/focus-flexbox-container';
 import canFocusImgIsmap from '../supports/focus-img-ismap';
 import canFocusImgUsemapTabindex from '../supports/focus-img-usemap-tabindex';
+import canFocusInHiddenIframe from '../supports/focus-in-hidden-iframe';
+import canFocusInZeroDimensionObject from '../supports/focus-in-zero-dimension-object';
 import canFocusLabelTabindex from '../supports/focus-label-tabindex';
 import canFocusObjectSvg from '../supports/focus-object-svg';
+import canFocusObjectSvgHidden from '../supports/focus-object-svg-hidden';
 import canFocusObjectSwf from '../supports/focus-object-swf';
 import canFocusRedirectImgUsemap from '../supports/focus-redirect-img-usemap';
 import canFocusRedirectLegend from '../supports/focus-redirect-legend';
@@ -29,8 +32,11 @@ export default memorizeResult(function() {
     canFocusFlexboxContainer: canFocusFlexboxContainer(),
     canFocusImgIsmap: canFocusImgIsmap(),
     canFocusImgUsemapTabindex: canFocusImgUsemapTabindex(),
+    canFocusInHiddenIframe: canFocusInHiddenIframe(),
+    canFocusInZeroDimensionObject: canFocusInZeroDimensionObject(),
     canFocusLabelTabindex: canFocusLabelTabindex(),
     canFocusObjectSvg: canFocusObjectSvg(),
+    canFocusObjectSvgHidden: canFocusObjectSvgHidden(),
     canFocusObjectSwf: canFocusObjectSwf(),
     canFocusRedirectImgUsemap: canFocusRedirectImgUsemap(),
     canFocusRedirectLegend: canFocusRedirectLegend(),