+++ src/is/focusable.js
@@ -14,6 +14,7 @@ import isVisible from './visible';
 import isDisabled from './disabled';
 import isOnlyTabbable from './only-tabbable';
 import contextToElement from '../util/context-to-element';
+import getFrameElement from '../util/get-frame-element';
 import getWindow from '../util/get-window';
 import tabindexValue from '../util/tabindex-value';
 
@@ -82,6 +83,11 @@ function isFocusableRules({
     supports = _supports();
   }
 
+  const _isOnlyTabbable = isOnlyTabbable.rules.except({
+    onlyFocusableBrowsingContext: true,
+    visible: except.visible,
+  });
+
   const element = contextToElement({
     message: 'is/focusable requires an argument of type Element',
     resolveDocument: true,
@@ -101,14 +107,46 @@ function isFocusableRules({
     return false;
   }
 
-  if (!except.onlyTabbable && isOnlyTabbable(element)) {
+  if (!except.onlyTabbable && _isOnlyTabbable(element)) {
     // some elements may be keyboard focusable, but not script focusable
     return false;
   }
 
   // elements that are not rendered, cannot be focused
-  if (!except.visible && !isVisible(element)) {
-    return false;
+  if (!except.visible) {
+    const visibilityOptions = {
+      context: element,
+      except: {},
+    };
+
+    if (supports.canFocusInHiddenIframe) {
+      // WebKit and Blink can focus content in hidden <iframe> and <object>
+      visibilityOptions.except.browsingContext = true;
+    }
+
+    if (supports.canFocusObjectSvgHidden) {
+      // Blink allows focusing the object element, even if it has visibility: hidden;
+      // @browser-issue Blink https://code.google.com/p/chromium/issues/detail?id=586191
+      const nodeName = element.nodeName.toLowerCase();
+      if (nodeName === 'object') {
+        visibilityOptions.except.cssVisibility = true;
+      }
+    }
+
+    if (!isVisible.rules(visibilityOptions)) {
+      return false;
+    }
+  }
+
+  const frameElement = getFrameElement(element);
+  if (frameElement) {
+    const _nodeName = frameElement.nodeName.toLowerCase();
+    if (_nodeName === 'object' && !supports.canFocusInZeroDimensionObject) {
+      if (!frameElement.offsetWidth || !frameElement.offsetHeight) {
+        // WebKit can not focus content in <object> if it doesn't have dimensions
+        return false;
+      }
+    }
   }
 
   return true;