+++ src/is/only-tabbable.js
@@ -1,15 +1,38 @@
 
+import isVisible from './visible';
 import contextToElement from '../util/context-to-element';
+import getFrameElement from '../util/get-frame-element';
 import getWindow from '../util/get-window';
 import tabindexValue from '../util/tabindex-value';
 import platform from '../util/platform';
 
-function isOnlyTabbableRules({ context } = {}) {
+function isOnlyTabbableRules({
+  context,
+  except = {
+    onlyFocusableBrowsingContext: false,
+    visible: false,
+  },
+} = {}) {
   const element = contextToElement({
     message: 'is/only-tabbable requires an argument of type Element',
     resolveDocument: true,
     context,
   });
+
+  if (!except.visible && !isVisible(element)) {
+    return false;
+  }
+
+  if (!except.onlyFocusableBrowsingContext && (platform.is.GECKO || platform.is.TRIDENT)) {
+    const frameElement = getFrameElement(element);
+    if (frameElement) {
+      if (tabindexValue(frameElement) < 0) {
+        // iframe[tabindex="-1"] and object[tabindex="-1"] inherit the
+        // tabbable demotion onto elements of their browsing contexts
+        return false;
+      }
+    }
+  }
 
   const nodeName = element.nodeName.toLowerCase();
   const tabindex = tabindexValue(element);