+++ src/supports/focus-in-zero-dimension-object.js
@@ -1,0 +1,8 @@
+
+import platform from '../util/platform';
+
+const result = !platform.is.WEBKIT;
+
+export default function() {
+  return result;
+}