+++ test/unit/get.parents.test.js
@@ -17,6 +17,11 @@ define([
           '<div>',
             '<span id="target">target</span>',
           '</div>',
+          '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" id="svg">',
+            '<a xlink:href="#void" id="svg-link">',
+              '<text x="10" y="20" id="svg-link-text">text</text>',
+            '</a>',
+          '</svg>',
           /*eslint-disable indent */
         ]);
       },
@@ -34,6 +39,16 @@ define([
         var expected = '#target div #intern-dom-fixture body html'.split(' ');
         var target = getParents({
           context: '#target',
+        });
+        var path = target.map(function(element) {
+          return element.id && ('#' + element.id) || element.nodeName.toLowerCase();
+        });
+        expect(path).to.deep.equal(expected);
+      },
+      'parents in SVG': function() {
+        var expected = '#svg-link-text #svg-link #svg #intern-dom-fixture body html'.split(' ');
+        var target = getParents({
+          context: '#svg-link-text',
         });
         var path = target.map(function(element) {
           return element.id && ('#' + element.id) || element.nodeName.toLowerCase();