+++ test/unit/is.visible.test.js
@@ -3,15 +3,40 @@ define([
   'intern/chai!expect',
   '../helper/fixtures/custom.fixture',
   '../helper/supports',
+  '../helper/test-frame',
   'ally/is/visible',
   'ally/supports/media/mp3',
-], function(registerSuite, expect, customFixture, supports, isVisible, mp3) {
+], function(registerSuite, expect, customFixture, supports, TestFrame, isVisible, mp3) {
 
   registerSuite(function() {
     var fixture;
+    var frame;
 
     return {
       name: 'is/visible',
+
+      before: function() {
+        frame = new TestFrame([
+          /*eslint-disable indent */
+          '<!DOCTYPE html>',
+          '<html lang="en">',
+            '<head>',
+              '<meta charset="utf-8" />',
+              '<title>Framed Content</title>',
+            '</head>',
+            '<body>',
+              '<p id="target">Hello World</p>',
+            '</body>',
+          '</html>',
+          /*eslint-enable indent */
+        ].join(''));
+
+        return frame.initialize(document.body);
+      },
+      after: function() {
+        frame.terminate();
+        frame = null;
+      },
 
       beforeEach: function() {
         fixture = customFixture([
@@ -70,6 +95,13 @@ define([
         expect(function() {
           isVisible(null);
         }).to.throw(TypeError, 'is/visible requires an argument of type Element');
+      },
+      '.rules() and .except()': function() {
+        var element = document.getElementById('visible-div');
+        expect(isVisible.rules({
+          context: element,
+        })).to.equal(true, '.rules()');
+        expect(isVisible.rules.except({})(element)).to.equal(true, '.rules.except()');
       },
       div: function() {
         var element = document.getElementById('visible-div');
@@ -149,6 +181,16 @@ define([
       'unknown dimension: audio element': function() {
         var element = document.getElementById('unknown-dimension-audio');
         expect(isVisible(element)).to.equal(true);
+      },
+      'within visible <iframe>': function() {
+        var element = frame.document.getElementById('target');
+        frame.element.style.display = '';
+        expect(isVisible(element)).to.equal(true);
+      },
+      'within hidden <iframe>': function() {
+        var element = frame.document.getElementById('target');
+        frame.element.style.display = 'none';
+        expect(isVisible(element)).to.equal(false);
       },
     };
   });