+++ tests/focusable/focusable.js
@@ -5,6 +5,7 @@ require.config({
     // shims required by ally.js
     'array.prototype.findindex': '../../node_modules/array.prototype.findindex/index',
     'css.escape': '../../node_modules/css.escape/css.escape',
+    'domtokenlist-shim': '../../node_modules/domtokenlist-shim/dist/domtokenlist',
     'es6-promise': '../../node_modules/es6-promise/dist/es6-promise',
     // stuff used for testing and co
     platform: '../../node_modules/platform/platform',
@@ -57,6 +58,17 @@ require([
   'es6-promise',
 ], function(Promise) {
   Promise.polyfill();
+
+  if (location.search.indexOf('manual=1') !== -1) {
+    require([
+      'platform',
+      'ally/ally',
+    ], function(platform, ally) {
+      window.platform = platform;
+      window.ally = ally;
+    });
+    return;
+  }
 
   if (window.isLoaded) {
     runFocusTests();